Deleted scene\exporter, fixed missing code references
