--- fast64_internal/oot/scene/exporter/to_c/scene_collision.py
@@ -1,9 +1,0 @@
-from ....collision.exporter.to_c import ootCollisionToC
-from ....oot_level_classes import OOTScene
-
-
-# Writes the collision data for a scene
-def getSceneCollision(outScene: OOTScene):
-    # @TODO: delete this function and rename ``ootCollisionToC`` into ``getSceneCollision``
-    # when the ``oot_collision.py`` code is cleaned up
-    return ootCollisionToC(outScene.collision)
